+++ com.unity.shadergraph/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Version Updated
 The version number for this package has increased due to a version update of a related graphics package.
 
+### Fixed
+- Fixed issue where a NullReferenceException would be thrown on resetting reference name for a Shader Graph property
+- 
+
 ## [10.2.0] - 2020-10-19
 
 ### Added